Remove button to create new wallet using web app
Use of the web app is now deprecated so modify the page that you see
when clicking on 'Create your Wallet' from the main web page to only
offer the links to download the standalone/native apps. It's still
possible to create a new wallet using the web app by navigating directly
to the url but we want to discourage it.