Re-enable settings for locked down wallets
There are several options in settings that should be available for
locked down wallets including:

PGP settings: if you lose your PGP key while in locked down mode you
need settings enabled to be able to disable it to cancel 2fa reset.
PIN settings: setting the PIN is allowed on the server in locked down
mode and you can set the PIN during login, but settings is missing which
is inconsistent.
Disable PIN: there is no way to disable PIN login without settings, and
this should be possible in locked down mode.